[BE] fixing update / delete transaction logic

Approved-by: Jong man Park <jongmanpark418@gmail.com>

+++ backend/src/main/java/com/ez/ezbackend/budget/service/TransactionServiceImpl.java
@@ -4,6 +4,7 @@ import com.ez.ezbackend.budget.entity.Transaction;
 import com.ez.ezbackend.budget.model.TransactionModel;
 import com.ez.ezbackend.budget.repository.TransactionRepository;
 import com.ez.ezbackend.shared.entity.User;
+import com.ez.ezbackend.shared.exception.EzNotAuthorizedException;
 import com.ez.ezbackend.shared.exception.EzNotFoundException;
 import com.ez.ezbackend.shared.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -42,10 +43,13 @@ public class TransactionServiceImpl implements TransactionService {
   public Transaction updateTransactionForUser(TransactionModel transactionRequest, long transactionId, long userId) {
     User user = userRepository.findById(userId)
         .orElseThrow(() -> new EzNotFoundException("User with ID: " + userId + " not found."));
-    transactionRepository.findById(transactionId)
+    Transaction transaction = transactionRepository.findById(transactionId)
         .orElseThrow(() -> new EzNotFoundException("Transaction with ID: " + transactionId + " not found."));
-    Transaction transaction = TransactionModel.convertToTransaction(transactionRequest, user, transactionId);
-    return transactionRepository.saveAndFlush(transaction);
+    if (userId != transaction.getUser().getId()) {
+      throw new EzNotAuthorizedException("User with ID: " + userId + " not authorized to delete the requested transaction.");
+    }
+    Transaction updatedTransaction = TransactionModel.convertToTransaction(transactionRequest, user, transactionId);
+    return transactionRepository.saveAndFlush(updatedTransaction);
   }
 
   @Override
@@ -54,6 +58,9 @@ public class TransactionServiceImpl implements TransactionService {
         .orElseThrow(() -> new EzNotFoundException("User with ID: " + userId + " not found."));
     Transaction transaction = transactionRepository.findById(transactionId)
         .orElseThrow(() -> new EzNotFoundException("Transaction with ID: " + transactionId + " not found."));
+    if (userId != transaction.getUser().getId()) {
+      throw new EzNotAuthorizedException("User with ID: " + userId + " not authorized to delete the requested transaction.");
+    }
     transactionRepository.delete(transaction);
   }
 }
